refactor: fixed compilation errors

+++ packages/polaris-typeorm/test/integration-tests/postgres-tests/polaris-repository.test.ts
@@ -35,13 +35,13 @@ let libraryRepo: PolarisRepository<Library>;
 
 beforeEach(async () => {
     connection = await setUpTestConnection();
-    authorRepo = connection.getRepository(Author);
-    bookRepo = connection.getRepository(Book);
-    profileRepo = connection.getRepository(Profile);
-    userRepo = connection.getRepository(User);
-    dvRepo = connection.getRepository(DataVersion);
-    libraryRepo = connection.getRepository(Library);
-    cookbookRepo = connection.getRepository(Cookbook);
+    authorRepo = connection.getRepository(Author, generateContext());
+    bookRepo = connection.getRepository(Book, generateContext());
+    profileRepo = connection.getRepository(Profile, generateContext());
+    userRepo = connection.getRepository(User, generateContext());
+    dvRepo = connection.getRepository(DataVersion, generateContext());
+    libraryRepo = connection.getRepository(Library, generateContext());
+    cookbookRepo = connection.getRepository(Cookbook, generateContext());
     await initDb(connection);
 });
 afterEach(async () => {
@@ -53,15 +53,14 @@ describe('entity manager tests', () => {
         it('parent is not common model, hard delete parent entity', async () => {
             const findConditions = { name: 'public' };
             const findOptions = { where: findConditions };
-            libraryRepo = connection.getRepository(Library, generateContext());
             await libraryRepo.delete(findConditions);
             const libAfterDelete = await libraryRepo.findOne(findOptions);
             expect(libAfterDelete).toBeUndefined();
         });
 
         it('field is not common model, does not delete linked entity', async () => {
-            await authorRepo.delete(generateContext(), authorWithCascadeFindOneOptions.where);
-            const lib = await libraryRepo.findOne(generateContext(), {
+            await authorRepo.delete(authorWithCascadeFindOneOptions.where);
+            const lib = await libraryRepo.findOne({
                 relations: ['books'],
             });
             const criteria = {
@@ -70,7 +69,7 @@ describe('entity manager tests', () => {
                     deleted: In([true, false]),
                 },
             };
-            const authorWithCascade = await authorRepo.findOne(generateContext(), criteria);
+            const authorWithCascade = await authorRepo.findOne(criteria);
             expect(lib).toBeDefined();
             authorWithCascade
                 ? expect(authorWithCascade.getDeleted()).toBeTruthy()
@@ -82,8 +81,8 @@ describe('entity manager tests', () => {
                 where: { ...userFindOneOptions.where, deleted: In([true, false]) },
                 relations: ['profile'],
             };
-            await userRepo.delete(generateContext(), criteria.where);
-            const userCommonModel = await userRepo.findOne(generateContext(), criteria);
+            await userRepo.delete(criteria.where);
+            const userCommonModel = await userRepo.findOne(criteria);
             userCommonModel
                 ? expect(userCommonModel.getDeleted()).toBeTruthy()
                 : expect(userCommonModel).toBeDefined();
@@ -108,15 +107,11 @@ describe('entity manager tests', () => {
                     deleted: In([true, false]),
                 },
             };
-            await authorRepo.delete(generateContext(), authorFindOneOptions1.where);
+            await authorRepo.delete(authorFindOneOptions1.where);
             const authorWithCascade: Author | undefined = await authorRepo.findOne(
-                generateContext(),
                 authorFindOneOptions1,
             );
-            const bookWithCascade: Book | undefined = await bookRepo.findOne(
-                generateContext(),
-                bookFindOneOptions1,
-            );
+            const bookWithCascade: Book | undefined = await bookRepo.findOne(bookFindOneOptions1);
             bookWithCascade
                 ? expect(bookWithCascade.getDeleted()).toBeTruthy()
                 : expect(bookWithCascade).toBeDefined();
@@ -126,8 +121,8 @@ describe('entity manager tests', () => {
         });
 
         it('delete linked entity, should not return deleted entities(first level), get entity and its linked entity', async () => {
-            await profileRepo.delete(generateContext(), profileFindOneOptions.where);
-            const userEntity: User | undefined = await userRepo.findOne(generateContext(), {
+            await profileRepo.delete(profileFindOneOptions.where);
+            const userEntity: User | undefined = await userRepo.findOne({
                 ...userFindOneOptions,
                 relations: ['profile'],
             });
@@ -141,11 +136,8 @@ describe('entity manager tests', () => {
 
         // checks default setting
         it('delete entity, should not return deleted entities, doesnt return deleted entity', async () => {
-            await bookRepo.delete(generateContext(), bookFindOneOptions.where);
-            const book: Book | undefined = await bookRepo.findOne(
-                generateContext(),
-                bookFindOneOptions,
-            );
+            await bookRepo.delete(bookFindOneOptions.where);
+            const book: Book | undefined = await bookRepo.findOne(bookFindOneOptions);
             expect(book).toBeUndefined();
         });
 
@@ -154,8 +146,8 @@ describe('entity manager tests', () => {
             Object.assign(connection.options, {
                 extra: { config: { allowSoftDelete: false } },
             });
-            await authorRepo.delete(generateContext(), authorFindOneOptions.where);
-            const author: Author | undefined = await authorRepo.findOne(generateContext(), {
+            await authorRepo.delete(authorFindOneOptions.where);
+            const author: Author | undefined = await authorRepo.findOne({
                 where: {
                     ...authorFindOneOptions.where,
                     deleted: In([true, false]),
@@ -173,14 +165,14 @@ describe('entity manager tests', () => {
                 Object.assign(connection.options, {
                     extra: { config: { allowSoftDelete: false } },
                 });
-                await authorRepo.delete(generateContext(), authorWithCascadeFindOneOptions.where);
-                const bookWithCascade = await bookRepo.findOne(generateContext(), {
+                await authorRepo.delete(authorWithCascadeFindOneOptions.where);
+                const bookWithCascade = await bookRepo.findOne({
                     where: {
                         ...bookWithCascadeFindOneOptions.where,
                         deleted: In([true, false]),
                     },
                 });
-                const authorWithCascade = await authorRepo.findOne(generateContext(), {
+                const authorWithCascade = await authorRepo.findOne({
                     where: {
                         ...authorWithCascadeFindOneOptions.where,
                         deleted: In([true, false]),
@@ -194,11 +186,9 @@ describe('entity manager tests', () => {
     });
     describe('data version tests', () => {
         it('books are created with data version, get all book for data version 0', async () => {
-            const booksInit = await bookRepo.find(generateContext(), {});
-            const booksAfterDataVersion = await bookRepo.find(
-                generateContext({ dataVersion: 2 }),
-                {},
-            );
+            const booksInit = await bookRepo.find({});
+            bookRepo = connection.getRepository(Book, generateContext({ dataVersion: 2 }));
+            const booksAfterDataVersion = await bookRepo.find({});
             expect(booksInit.length).toEqual(2);
             expect(booksAfterDataVersion.length).toEqual(0);
         });
@@ -207,17 +197,16 @@ describe('entity manager tests', () => {
         it('reality id is supplied in headers', async () => {
             const bookReality1: any = new Book('Jurassic Park');
             bookReality1.realityId = 1;
-            await bookRepo.save(generateContext({ realityId: 1 }), bookReality1);
-            const book = await bookRepo.findOne(generateContext({ realityId: 1 }), {});
+            bookRepo = connection.getRepository(Book, generateContext({ realityId: 1 }));
+            await bookRepo.save(bookReality1);
+            const book = await bookRepo.findOne({});
             expect(book).toEqual(bookReality1);
         });
 
-        it('delete operational entity, linked oper header true and reality id isnt operational, entity not deleted', async () => {
+        it('delete operational entity, linked oper header true and reality id is not operational, entity not deleted', async () => {
+            bookRepo = connection.getRepository(Book, generateContext({ realityId: 1 }));
             try {
-                await authorRepo.delete(
-                    generateContext({ realityId: 1 }),
-                    authorFindOneOptions.where,
-                );
+                await authorRepo.delete(authorFindOneOptions.where);
             } catch (err) {
                 expect(err.message).toEqual('there are no entities to delete');
             }
@@ -225,10 +214,11 @@ describe('entity manager tests', () => {
 
         it('save existing entity with different reality id, fail saving', async () => {
             const book: any = new Book('my book');
-            await bookRepo.save(generateContext(), book);
+            await bookRepo.save(book);
             book.realityId = 1;
+            bookRepo = connection.getRepository(Book, generateContext({ realityId: 1 }));
             try {
-                await bookRepo.save(generateContext({ realityId: 1 }), book);
+                await bookRepo.save(book);
             } catch (err) {
                 expect(err.message).toEqual('reality id of entity is different from header');
             }
@@ -236,17 +226,17 @@ describe('entity manager tests', () => {
     });
     it('find one with id', async () => {
         const book = new Book('my book');
-        await bookRepo.save(generateContext(), book);
-        const bookFound = await bookRepo.findOne(generateContext(), {
+        await bookRepo.save(book);
+        const bookFound = await bookRepo.findOne({
             where: { id: book.getId() },
         });
         expect(book).toEqual(bookFound);
     });
     it('count', async () => {
-        expect(await bookRepo.count(generateContext(), {})).toEqual(2);
+        expect(await bookRepo.count({})).toEqual(2);
     });
     it('order by', async () => {
-        const books1 = await bookRepo.find(generateContext(), {
+        const books1 = await bookRepo.find({
             order: {
                 title: 'ASC',
             },
@@ -258,12 +248,13 @@ describe('entity manager tests', () => {
         const book = new Book('my book');
 
         const createdByUpn = 'foo';
-        await bookRepo.save(generateContext({ upn: createdByUpn }), book);
+        bookRepo = connection.getRepository(Book, generateContext({ upn: createdByUpn }));
+        await bookRepo.save(book);
         expect(book.getCreatedBy()).toBe(createdByUpn);
         expect(book.getLastUpdatedBy()).toBe(createdByUpn);
 
         const updatedByUpn = 'bar';
-        await bookRepo.save(generateContext({ upn: updatedByUpn }), book);
+        await bookRepo.save(book);
         expect(book.getCreatedBy()).not.toBe(updatedByUpn);
         expect(book.getLastUpdatedBy()).toBe(updatedByUpn);
     });
@@ -271,15 +262,17 @@ describe('entity manager tests', () => {
         const book = new Book('my book');
 
         const createdByUpn = 'foo';
+        bookRepo = connection.getRepository(Book, generateContext({ upn: createdByUpn }));
         book.setCreationTime(new Date());
-        await bookRepo.save(generateContext({ upn: createdByUpn }), book);
+        await bookRepo.save(book);
         expect(book.getCreatedBy()).toBe(createdByUpn);
         expect(book.getLastUpdatedBy()).toBe(createdByUpn);
     });
     it('create grandchild entity with upn in context, upn is set to entity', async () => {
         const cookbook = new Cookbook('tasty food');
         const upn = 'foo';
-        await cookbookRepo.save(generateContext({ upn }), cookbook);
+        bookRepo = connection.getRepository(Book, generateContext({ upn }));
+        await cookbookRepo.save(cookbook);
         expect(cookbook.getCreatedBy()).toBe(upn);
     });
     it('save entity with upn, entity already has creation time, creation time is not overridden', async () => {
@@ -287,14 +280,18 @@ describe('entity manager tests', () => {
 
         const createdByUpn = 'foo';
         book.setCreationTime(new Date());
-        await bookRepo.save(generateContext({ upn: createdByUpn }), book);
-        const bookFound = await bookRepo.findOne(generateContext(), {
+        const bookRepoWithHeaders = connection.getRepository(
+            Book,
+            generateContext({ upn: createdByUpn }),
+        );
+        await bookRepoWithHeaders.save(book);
+        const bookFound = await bookRepo.findOne({
             where: { id: book.getId() },
         });
         const creationDate = bookFound?.getCreationTime();
         book.setCreationTime(new Date(1));
-        await bookRepo.save(generateContext({ upn: createdByUpn }), book);
-        const bookFoundAfterCreationTimeChange = await bookRepo.findOne(generateContext(), {
+        await bookRepoWithHeaders.save(book);
+        const bookFoundAfterCreationTimeChange = await bookRepo.findOne({
             where: { id: book.getId() },
         });
         expect(creationDate).not.toStrictEqual(bookFoundAfterCreationTimeChange?.getCreationTime());
@@ -302,53 +299,57 @@ describe('entity manager tests', () => {
     describe('find by ids tests', () => {
         it('create entity find it by ids, returns entity', async () => {
             const book = new Book('my book');
-            await bookRepo.save(generateContext(), book);
-            const bookFound = await bookRepo.findByIds(generateContext({ realityId: 0 }), [
-                book.getId(),
-            ]);
+            const bookRepoWithHeaders = connection.getRepository(
+                Book,
+                generateContext({ realityId: 0 }),
+            );
+            await bookRepo.save(book);
+            const bookFound = await bookRepoWithHeaders.findByIds([book.getId()]);
             expect(bookFound[0]).toEqual(book);
         });
         it('delete entity then find it by ids, returns entity', async () => {
             const book = new Book('my book');
-            await bookRepo.save(generateContext(), book);
-            await bookRepo.delete(generateContext(), book.getId());
-            const bookFound = await bookRepo.findByIds(generateContext({ realityId: 0 }), [
-                book.getId(),
-            ]);
+            await bookRepo.save(book);
+            await bookRepo.delete(book.getId());
+            bookRepo = connection.getRepository(Book, generateContext({ realityId: 0 }));
+            const bookFound = await bookRepo.findByIds([book.getId()]);
             expect(bookFound[0]).toBeDefined();
         });
     });
     describe('query builder tests', () => {
         it('find one with id', async () => {
             const book = new Book('my book');
-            await bookRepo.save(generateContext(), book);
+            await bookRepo.save(book);
+            bookRepo = connection.getRepository(
+                Book,
+                generateContext({ realityId: 0, dataVersion: book.getDataVersion() - 1 }),
+            );
             const bookFound = await bookRepo
-                .createQueryBuilder(
-                    generateContext({ realityId: 0, dataVersion: book.getDataVersion() - 1 }),
-                    'book',
-                )
+                .createQueryBuilder('book')
                 .andWhere('book.id = :id ', { id: book.getId() })
                 .getOne();
             expect(bookFound?.getId()).toEqual(book.getId());
         });
         it('find one with data version equal to the entity, returns no entity', async () => {
             const book = new Book('my book');
-            await bookRepo.save(generateContext(), book);
+            await bookRepo.save(book);
+            bookRepo = connection.getRepository(
+                Book,
+                generateContext({ realityId: 0, dataVersion: book.getDataVersion() }),
+            );
             const bookFound = await bookRepo
-                .createQueryBuilder(
-                    generateContext({ realityId: 0, dataVersion: book.getDataVersion() }),
-                    'book',
-                )
+                .createQueryBuilder('book')
                 .andWhere('book.id = :id ', { id: book.getId() })
                 .getOne();
             expect(bookFound).toBeUndefined();
         });
         it('find deleted entity, returns no entity', async () => {
             const book = new Book('my book');
-            await bookRepo.save(generateContext(), book);
-            await bookRepo.delete(generateContext(), book.getId());
+            await bookRepo.save(book);
+            await bookRepo.delete(book.getId());
+            bookRepo = connection.getRepository(Book, generateContext({ realityId: 0 }));
             const bookFound = await bookRepo
-                .createQueryBuilder(generateContext({ realityId: 0 }), 'book')
+                .createQueryBuilder('book')
                 .andWhere('book.id = :id ', { id: book.getId() })
                 .getOne();
             expect(bookFound).toBeUndefined();
